Make inflected column name to be in natural case instead of lowercase
The 'inflected_name' property of column metadata must be in the same letter case as column names seen in a response from a mysql server to work properly.

The column names also must be in the natural case to make it compatible with mysql_* functions responses and to make model attributes in the phpactiverecord usage also in the natural case.

To make it happen:

* StandardInflector->variablize($s) no more lowercases column names.
* Connection::$PDO_OPTIONS[PDO::ATTR_CASE] is set to PDO::CASE_NATURAL.

After this change all usage of model attributes that correspond to columns having uppercase letters must be used as-is now.

Example:
Table `commutators` have a column `networkNode_id`.

Before this change it should be written as:
$commutator->networknode_id

After this change it should be written as:
$commutator->networkNode_id